GetById implementation with unit tests
