+++ Apsitvarkom.DataAccess/PollutedLocationDTOFileRepository.cs
@@ -7,14 +7,14 @@ namespace Apsitvarkom.DataAccess;
 /// <summary>
 /// Class for <see cref="PollutedLocationDTO" /> data handling from file.
 /// </summary>
-public class PollutedLocationsDTOFileRepository : IPollutedLocationDTORepository, IDisposable
+public class PollutedLocationDTOFileRepository : IPollutedLocationDTORepository, IDisposable
 {
     private readonly JsonSerializerOptions _options;
     private readonly Stream _stream;
 
     /// <summary>Constructor for the reader.</summary>
     /// <param name="stream">Stream to be used for parsing.</param>
-    private PollutedLocationsDTOFileRepository(Stream stream)
+    private PollutedLocationDTOFileRepository(Stream stream)
     {
         _stream = stream;
         _options = new JsonSerializerOptions
@@ -25,27 +25,34 @@ public class PollutedLocationsDTOFileRepository : IPollutedLocationDTORepository
     }
 
     /// <inheritdoc />
-    public async Task<IEnumerable<PollutedLocationDTO>> GetAllPollutedLocationsAsync()
+    public async Task<IEnumerable<PollutedLocationDTO>> GetAllAsync()
     {
         var result = await JsonSerializer.DeserializeAsync<IEnumerable<PollutedLocationDTO>>(_stream, _options);
         return result ?? Enumerable.Empty<PollutedLocationDTO>();
     }
 
+    /// <inheritdoc />
+    public async Task<PollutedLocationDTO?> GetByIdAsync(string id)
+    {
+        var result = await GetAllAsync();
+        return result.SingleOrDefault(loc => loc.Id == id); 
+    }
+
     /// <summary>Static factory constructor for reader from file.</summary>
     /// <param name="sourcePath">Relative location of the .json type source data file.</param>
-    public static PollutedLocationsDTOFileRepository FromFile(string sourcePath)
+    public static PollutedLocationDTOFileRepository FromFile(string sourcePath)
     {
         var stream = File.OpenRead(sourcePath);
-        return new PollutedLocationsDTOFileRepository(stream);
+        return new PollutedLocationDTOFileRepository(stream);
     }
 
     /// <summary>Static factory constructor for reader from JSON string.</summary>
     /// <param name="contents">Contents to be parsed from JSON string.</param>
-    public static PollutedLocationsDTOFileRepository FromContent(string contents)
+    public static PollutedLocationDTOFileRepository FromContent(string contents)
     {
         var byteArray = Encoding.UTF8.GetBytes(contents);
         var stream = new MemoryStream(byteArray);
-        return new PollutedLocationsDTOFileRepository(stream);
+        return new PollutedLocationDTOFileRepository(stream);
     }
 
     /// <inheritdoc />
